integrate scroll area module with the scroll command
