+++ src/index.js
@@ -16,7 +16,6 @@ import open from './open';
 import pickLabel from './pick-label';
 import rangeOptions from './range-options';
 import scroll from './scroll';
-import scrollArea from './scroll-area';
 import search from './search';
 import select from './select';
 import selectOptions from './select-options';
@@ -44,7 +43,6 @@ const coreModules = [
   rangeOptions,
   search,
   scroll,
-  scrollArea,
   select,
   selectOptions,
   textOptions,
